Refactor cloud sync state ownership
