+++ src/utils/cloudSync.ts
@@ -1,16 +1,10 @@
 import { abortableFetch } from "@/utils/abortableFetch";
 import { ensureIndexedDBInitialized, STORES } from "@/utils/indexedDB";
 import { getApiUrl } from "@/utils/platform";
-import { useThemeStore } from "@/stores/useThemeStore";
-import { useLanguageStore } from "@/stores/useLanguageStore";
 import { useDisplaySettingsStore } from "@/stores/useDisplaySettingsStore";
-import { useAudioSettingsStore } from "@/stores/useAudioSettingsStore";
-import { useAppStore } from "@/stores/useAppStore";
 import { useFilesStore, type FileSystemItem } from "@/stores/useFilesStore";
 import { useIpodStore, type Track } from "@/stores/useIpodStore";
 import { useVideoStore, type Video } from "@/stores/useVideoStore";
-import { useDockStore } from "@/stores/useDockStore";
-import { useDashboardStore } from "@/stores/useDashboardStore";
 import { useStickiesStore, type StickyNote } from "@/stores/useStickiesStore";
 import {
   useCalendarStore,
@@ -18,13 +12,10 @@ import {
   type CalendarGroup,
   type TodoItem,
 } from "@/stores/useCalendarStore";
-import { useContactsStore } from "@/stores/useContactsStore";
 import {
   useCloudSyncStore,
   type CloudSyncDeletionBucket,
 } from "@/stores/useCloudSyncStore";
-import type { Contact } from "@/utils/contacts";
-import { normalizeContacts } from "@/utils/contacts";
 import {
   uploadBlobWithStorageInstruction,
 } from "@/utils/storageUpload";
@@ -48,7 +39,7 @@ import {
 import {
   getNextSyncClientVersion,
   getSyncClientId,
-} from "@/utils/cloudSyncClientState";
+} from "@/utils/sync/engine/state/syncStateAdapter";
 import { getSyncSessionId } from "@/utils/syncSession";
 import {
   fetchBlobDomainPayload,
@@ -67,23 +58,24 @@ import {
   type FilesMetadataSyncSnapshot,
 } from "@/utils/cloudSyncFileMerge";
 import {
-  beginApplyingRemoteSettingsSections,
-  endApplyingRemoteSettingsSections,
-  getSettingsSectionTimestampMap,
-  setSettingsSectionTimestamps,
-  type SettingsSyncSection,
-} from "@/utils/cloudSyncSettingsState";
-import {
   beginApplyingRemoteDomain,
   endApplyingRemoteDomain,
 } from "@/utils/cloudSyncRemoteApplyState";
 import {
-  getRemoteSettingsSectionsToApply,
   mergeSettingsSnapshotData,
-  normalizeSettingsSnapshotData,
   shouldRestoreLegacyCustomWallpapers,
   type SettingsSnapshotData,
 } from "@/utils/cloudSyncSettingsMerge";
+import {
+  applyContactsSyncSnapshot,
+  mergeContactsSyncSnapshots,
+  serializeContactsSyncSnapshot,
+  type ContactsSnapshotData,
+} from "@/utils/sync/engine/domains/contactsSyncDomain";
+import {
+  applySettingsSyncSnapshot,
+  serializeSettingsSyncSnapshot,
+} from "@/utils/sync/engine/domains/settingsSyncDomain";
 import {
   getIndividualBlobKnownItems,
   setIndividualBlobKnownItems,
@@ -140,12 +132,6 @@ interface CalendarSnapshotData {
   deletedEventIds?: DeletionMarkerMap;
   deletedCalendarIds?: DeletionMarkerMap;
   deletedTodoIds?: DeletionMarkerMap;
-}
-
-interface ContactsSnapshotData {
-  contacts: Contact[];
-  myContactId: string | null;
-  deletedContactIds?: DeletionMarkerMap;
 }
 
 type AnySnapshotData =
@@ -549,68 +535,7 @@ async function gzipJson(value: unknown): Promise<Uint8Array> {
 }
 
 
-export function serializeSettingsSnapshot(): SettingsSnapshotData {
-  const displayState = useDisplaySettingsStore.getState();
-  const audioState = useAudioSettingsStore.getState();
-  const ipodState = useIpodStore.getState();
-  const dockState = useDockStore.getState();
-  const dashboardState = useDashboardStore.getState();
-  const sectionUpdatedAt = getSettingsSectionTimestampMap();
-
-  return {
-    theme: useThemeStore.getState().current,
-    language: useLanguageStore.getState().current,
-    languageInitialized:
-      localStorage.getItem("ryos:language-initialized") === "true",
-    aiModel: useAppStore.getState().aiModel,
-    display: {
-      displayMode: displayState.displayMode,
-      shaderEffectEnabled: displayState.shaderEffectEnabled,
-      selectedShaderType: displayState.selectedShaderType,
-      currentWallpaper: displayState.currentWallpaper,
-      screenSaverEnabled: displayState.screenSaverEnabled,
-      screenSaverType: displayState.screenSaverType,
-      screenSaverIdleTime: displayState.screenSaverIdleTime,
-      debugMode: displayState.debugMode,
-      htmlPreviewSplit: displayState.htmlPreviewSplit,
-    },
-    audio: {
-      masterVolume: audioState.masterVolume,
-      uiVolume: audioState.uiVolume,
-      chatSynthVolume: audioState.chatSynthVolume,
-      speechVolume: audioState.speechVolume,
-      ipodVolume: audioState.ipodVolume,
-      uiSoundsEnabled: audioState.uiSoundsEnabled,
-      terminalSoundsEnabled: audioState.terminalSoundsEnabled,
-      typingSynthEnabled: audioState.typingSynthEnabled,
-      speechEnabled: audioState.speechEnabled,
-      keepTalkingEnabled: audioState.keepTalkingEnabled,
-      ttsModel: audioState.ttsModel,
-      ttsVoice: audioState.ttsVoice,
-      synthPreset: audioState.synthPreset,
-    },
-    ipod: {
-      displayMode: ipodState.displayMode,
-      showLyrics: ipodState.showLyrics,
-      lyricsAlignment: ipodState.lyricsAlignment,
-      lyricsFont: ipodState.lyricsFont,
-      romanization: ipodState.romanization,
-      lyricsTranslationLanguage: ipodState.lyricsTranslationLanguage ?? null,
-      theme: ipodState.theme,
-      lcdFilterOn: ipodState.lcdFilterOn,
-    },
-    dock: {
-      pinnedItems: dockState.pinnedItems,
-      scale: dockState.scale,
-      hiding: dockState.hiding,
-      magnification: dockState.magnification,
-    },
-    dashboard: {
-      widgets: dashboardState.widgets,
-    },
-    sectionUpdatedAt,
-  };
-}
+export const serializeSettingsSnapshot = serializeSettingsSyncSnapshot;
 
 async function serializeCustomWallpapersSnapshot(): Promise<CustomWallpapersSnapshotData> {
   const db = await ensureIndexedDBInitialized();
@@ -796,14 +721,7 @@ function serializeCalendarSnapshot(): CalendarSnapshotData {
   };
 }
 
-function serializeContactsSnapshot(): ContactsSnapshotData {
-  const deletionMarkers = useCloudSyncStore.getState().deletionMarkers;
-  return {
-    contacts: useContactsStore.getState().contacts,
-    myContactId: useContactsStore.getState().myContactId,
-    deletedContactIds: deletionMarkers.contactIds,
-  };
-}
+const serializeContactsSnapshot = serializeContactsSyncSnapshot;
 
 export async function createCloudSyncEnvelope(
   domain: CloudSyncDomain
@@ -895,189 +813,43 @@ async function applySettingsSnapshot(
   data: SettingsSnapshotData,
   fallbackUpdatedAt: string
 ): Promise<void> {
-  const normalizedData = normalizeSettingsSnapshotData(data, fallbackUpdatedAt);
-  const remoteSectionUpdatedAt = normalizedData.sectionUpdatedAt || {};
-  const localSectionUpdatedAt = getSettingsSectionTimestampMap();
-  const sectionsToApply = getRemoteSettingsSectionsToApply(
-    localSectionUpdatedAt,
-    remoteSectionUpdatedAt
-  );
+  await applySettingsSyncSnapshot(data, fallbackUpdatedAt, {
+    beforeApply: async (normalizedData) => {
+      const legacyCustomWallpapers = normalizedData.customWallpapers || [];
+      const hasDedicatedCustomWallpaperSync = Boolean(
+        useCloudSyncStore.getState().remoteMetadata["custom-wallpapers"]?.updatedAt
+      );
 
-  const legacyCustomWallpapers = normalizedData.customWallpapers || [];
-  const hasDedicatedCustomWallpaperSync = Boolean(
-    useCloudSyncStore.getState().remoteMetadata["custom-wallpapers"]?.updatedAt
-  );
-
-  // Legacy: restore embedded custom wallpapers only on first migration when
-  // there is no dedicated custom-wallpapers sync domain and nothing local yet.
-  if (legacyCustomWallpapers.length > 0) {
-    const db = await ensureIndexedDBInitialized();
-    try {
-      const localWallpaperCount = (
-        await readStoreItems(db, STORES.CUSTOM_WALLPAPERS)
-      ).length;
-      if (
-        shouldRestoreLegacyCustomWallpapers({
-          legacyWallpaperCount: legacyCustomWallpapers.length,
-          localWallpaperCount,
-          hasDedicatedCustomWallpaperSync,
-        })
-      ) {
-        await upsertStoreItems(
-          db,
-          STORES.CUSTOM_WALLPAPERS,
-          legacyCustomWallpapers
-        );
-        useDisplaySettingsStore.getState().bumpCustomWallpapersRevision();
+      // Legacy: restore embedded custom wallpapers only on first migration when
+      // there is no dedicated custom-wallpapers sync domain and nothing local yet.
+      if (legacyCustomWallpapers.length === 0) {
+        return;
       }
-    } finally {
-      db.close();
-    }
-  }
 
-  const appliedSections: SettingsSyncSection[] = [];
-
-  beginApplyingRemoteSettingsSections(sectionsToApply);
-  try {
-    if (sectionsToApply.includes("theme")) {
+      const db = await ensureIndexedDBInitialized();
       try {
-        useThemeStore.getState().setTheme(normalizedData.theme as never);
-        appliedSections.push("theme");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote theme:", e);
+        const localWallpaperCount = (
+          await readStoreItems(db, STORES.CUSTOM_WALLPAPERS)
+        ).length;
+        if (
+          shouldRestoreLegacyCustomWallpapers({
+            legacyWallpaperCount: legacyCustomWallpapers.length,
+            localWallpaperCount,
+            hasDedicatedCustomWallpaperSync,
+          })
+        ) {
+          await upsertStoreItems(
+            db,
+            STORES.CUSTOM_WALLPAPERS,
+            legacyCustomWallpapers
+          );
+          useDisplaySettingsStore.getState().bumpCustomWallpapersRevision();
+        }
+      } finally {
+        db.close();
       }
-    }
-
-    if (sectionsToApply.includes("language")) {
-      try {
-        localStorage.setItem(
-          "ryos:language-initialized",
-          normalizedData.languageInitialized ? "true" : "false"
-        );
-        await useLanguageStore.getState().setLanguage(normalizedData.language);
-        appliedSections.push("language");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote language:", e);
-      }
-    }
-
-    if (sectionsToApply.includes("display")) {
-      try {
-        useDisplaySettingsStore.setState({
-          displayMode: normalizedData.display.displayMode as never,
-          shaderEffectEnabled: normalizedData.display.shaderEffectEnabled,
-          selectedShaderType: normalizedData.display.selectedShaderType as never,
-          screenSaverEnabled: normalizedData.display.screenSaverEnabled,
-          screenSaverType: normalizedData.display.screenSaverType,
-          screenSaverIdleTime: normalizedData.display.screenSaverIdleTime,
-          debugMode: normalizedData.display.debugMode,
-          htmlPreviewSplit: normalizedData.display.htmlPreviewSplit,
-        });
-
-        await useDisplaySettingsStore
-          .getState()
-          .setWallpaper(normalizedData.display.currentWallpaper);
-        appliedSections.push("display");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote display:", e);
-      }
-    }
-
-    if (sectionsToApply.includes("audio")) {
-      try {
-        useAudioSettingsStore.setState({
-          masterVolume: normalizedData.audio.masterVolume,
-          uiVolume: normalizedData.audio.uiVolume,
-          chatSynthVolume: normalizedData.audio.chatSynthVolume,
-          speechVolume: normalizedData.audio.speechVolume,
-          ipodVolume: normalizedData.audio.ipodVolume,
-          uiSoundsEnabled: normalizedData.audio.uiSoundsEnabled,
-          terminalSoundsEnabled: normalizedData.audio.terminalSoundsEnabled,
-          typingSynthEnabled: normalizedData.audio.typingSynthEnabled,
-          speechEnabled: normalizedData.audio.speechEnabled,
-          keepTalkingEnabled: normalizedData.audio.keepTalkingEnabled,
-          ttsModel: normalizedData.audio.ttsModel,
-          ttsVoice: normalizedData.audio.ttsVoice,
-          synthPreset: normalizedData.audio.synthPreset,
-        });
-        appliedSections.push("audio");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote audio:", e);
-      }
-    }
-
-    if (sectionsToApply.includes("aiModel")) {
-      try {
-        useAppStore.getState().setAiModel(normalizedData.aiModel);
-        appliedSections.push("aiModel");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote aiModel:", e);
-      }
-    }
-
-    if (normalizedData.ipod && sectionsToApply.includes("ipod")) {
-      try {
-        const remoteIpod = normalizedData.ipod;
-        useIpodStore.setState({
-          displayMode: remoteIpod.displayMode,
-          showLyrics: remoteIpod.showLyrics,
-          lyricsAlignment: remoteIpod.lyricsAlignment,
-          lyricsFont: remoteIpod.lyricsFont,
-          romanization: remoteIpod.romanization,
-          lyricsTranslationLanguage: remoteIpod.lyricsTranslationLanguage ?? null,
-          theme: remoteIpod.theme,
-          lcdFilterOn: remoteIpod.lcdFilterOn,
-        });
-        appliedSections.push("ipod");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote ipod:", e);
-      }
-    }
-
-    if (normalizedData.dock && sectionsToApply.includes("dock")) {
-      try {
-        useDockStore.setState({
-          pinnedItems: normalizedData.dock.pinnedItems,
-          scale: normalizedData.dock.scale,
-          hiding: normalizedData.dock.hiding,
-          magnification: normalizedData.dock.magnification,
-        });
-        appliedSections.push("dock");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote dock:", e);
-      }
-    }
-
-    if (
-      normalizedData.dashboard?.widgets &&
-      Array.isArray(normalizedData.dashboard.widgets) &&
-      sectionsToApply.includes("dashboard")
-    ) {
-      try {
-        useDashboardStore.setState({
-          widgets: normalizedData.dashboard.widgets,
-        });
-        appliedSections.push("dashboard");
-      } catch (e) {
-        console.error("[CloudSync] Failed to apply remote dashboard:", e);
-      }
-    }
-  } finally {
-    endApplyingRemoteSettingsSections(sectionsToApply);
-  }
-
-  if (appliedSections.length < sectionsToApply.length) {
-    const failed = sectionsToApply.filter((s) => !appliedSections.includes(s));
-    console.warn(
-      `[CloudSync] Settings apply: ${appliedSections.length}/${sectionsToApply.length} sections succeeded, failed: ${failed.join(", ")}`
-    );
-  }
-
-  setSettingsSectionTimestamps(
-    Object.fromEntries(
-      appliedSections.map((section) => [section, remoteSectionUpdatedAt[section] || fallbackUpdatedAt])
-    )
-  );
+    },
+  });
 }
 
 async function applyIndexedDbStoreSnapshot(
@@ -1212,29 +984,7 @@ function applyCalendarSnapshot(data: CalendarSnapshotData): void {
   });
 }
 
-function applyContactsSnapshot(data: ContactsSnapshotData): void {
-  const remoteDeletedContactIds = normalizeDeletionMarkerMap(
-    data.deletedContactIds
-  );
-  const cloudSyncState = useCloudSyncStore.getState();
-  const effectiveDeletedContactIds = mergeDeletionMarkerMaps(
-    cloudSyncState.deletionMarkers.contactIds,
-    remoteDeletedContactIds
-  );
-
-  cloudSyncState.mergeDeletedKeys("contactIds", remoteDeletedContactIds);
-
-  useContactsStore
-    .getState()
-    .replaceContactsFromSync(
-      filterDeletedIds(
-        normalizeContacts(data?.contacts),
-        effectiveDeletedContactIds,
-        (contact) => contact.id
-      ),
-      data?.myContactId ?? null
-    );
-}
+const applyContactsSnapshot = applyContactsSyncSnapshot;
 
 async function applyMonolithicBlobSnapshotToIndividualDomain(
   domain: IndividualBlobSyncDomain,
@@ -1516,24 +1266,7 @@ function mergeCalendarSnapshots(
   };
 }
 
-function mergeContactsSnapshots(
-  local: ContactsSnapshotData,
-  remote: ContactsSnapshotData
-): ContactsSnapshotData {
-  const mergedDeleted = mergeDeletionMarkerMaps(
-    normalizeDeletionMarkerMap(local.deletedContactIds),
-    normalizeDeletionMarkerMap(remote.deletedContactIds)
-  );
-  return {
-    contacts: mergeItemsByIdPreferNewer(
-      local.contacts,
-      normalizeContacts(remote.contacts),
-      mergedDeleted
-    ),
-    myContactId: local.myContactId ?? remote.myContactId,
-    deletedContactIds: mergedDeleted,
-  };
-}
+const mergeContactsSnapshots = mergeContactsSyncSnapshots;
 
 function mergeSongsSnapshots(
   local: SongsSnapshotData,
