+++ src/utils/syncLogicalDirtyState.ts
@@ -1,119 +1,67 @@
 import {
   getLogicalCloudSyncDomainForPhysical,
   getLogicalCloudSyncDomainPhysicalParts,
-  LOGICAL_CLOUD_SYNC_DOMAINS,
   type LogicalCloudSyncDomain,
 } from "@/utils/syncLogicalDomains";
+import { useCloudSyncStore } from "@/stores/useCloudSyncStore";
 import { type CloudSyncDomain } from "@/utils/cloudSyncShared";
-
-const STORAGE_KEY = "ryos:cloud-sync:logical-dirty-parts";
-
-type LogicalDirtyState = Partial<Record<LogicalCloudSyncDomain, CloudSyncDomain[]>>;
-
-function normalizeDirtyState(value: unknown): LogicalDirtyState {
-  if (!value || typeof value !== "object") {
-    return {};
-  }
-
-  const normalized: LogicalDirtyState = {};
-
-  for (const logicalDomain of LOGICAL_CLOUD_SYNC_DOMAINS) {
-    const rawParts = (value as Record<string, unknown>)[logicalDomain];
-    if (!Array.isArray(rawParts)) {
-      continue;
-    }
-
-    const allowedParts = new Set(
-      getLogicalCloudSyncDomainPhysicalParts(logicalDomain)
-    );
-    const uniqueParts = Array.from(
-      new Set(
-        rawParts.filter(
-          (part): part is CloudSyncDomain =>
-            typeof part === "string" && allowedParts.has(part as CloudSyncDomain)
-        )
-      )
-    );
-
-    if (uniqueParts.length > 0) {
-      normalized[logicalDomain] = uniqueParts;
-    }
-  }
-
-  return normalized;
-}
-
-function readDirtyState(): LogicalDirtyState {
-  if (typeof localStorage === "undefined") {
-    return {};
-  }
-
-  try {
-    const raw = localStorage.getItem(STORAGE_KEY);
-    if (!raw) {
-      return {};
-    }
-    return normalizeDirtyState(JSON.parse(raw));
-  } catch {
-    return {};
-  }
-}
-
-function writeDirtyState(state: LogicalDirtyState): void {
-  if (typeof localStorage === "undefined") {
-    return;
-  }
-
-  localStorage.setItem(STORAGE_KEY, JSON.stringify(state));
-}
 
 export function getPersistedLogicalDirtyParts(
   logicalDomain: LogicalCloudSyncDomain
 ): CloudSyncDomain[] {
-  return readDirtyState()[logicalDomain] || [];
+  return useCloudSyncStore.getState().persistedMetadata.logicalDirtyParts[logicalDomain] || [];
 }
 
 export function markLogicalDirtyPart(partDomain: CloudSyncDomain): void {
-  if (typeof localStorage === "undefined") {
-    return;
-  }
-
   const logicalDomain = getLogicalCloudSyncDomainForPhysical(partDomain);
-  const state = readDirtyState();
-  const nextParts = Array.from(
-    new Set([...(state[logicalDomain] || []), partDomain])
-  );
-  state[logicalDomain] = nextParts;
-  writeDirtyState(state);
+  useCloudSyncStore.getState().updatePersistedMetadata((current) => ({
+    ...current,
+    logicalDirtyParts: {
+      ...current.logicalDirtyParts,
+      [logicalDomain]: Array.from(
+        new Set([...(current.logicalDirtyParts[logicalDomain] || []), partDomain])
+      ),
+    },
+  }));
 }
 
 export function clearPersistedLogicalDirtyParts(
   logicalDomain: LogicalCloudSyncDomain,
   partDomains?: Iterable<CloudSyncDomain>
 ): void {
-  if (typeof localStorage === "undefined") {
-    return;
-  }
+  useCloudSyncStore.getState().updatePersistedMetadata((current) => {
+    const nextLogicalDirtyParts = {
+      ...current.logicalDirtyParts,
+    };
 
-  const state = readDirtyState();
+    if (!partDomains) {
+      delete nextLogicalDirtyParts[logicalDomain];
+      return {
+        ...current,
+        logicalDirtyParts: nextLogicalDirtyParts,
+      };
+    }
 
-  if (!partDomains) {
-    delete state[logicalDomain];
-    writeDirtyState(state);
-    return;
-  }
+    const allowedParts = new Set(
+      getLogicalCloudSyncDomainPhysicalParts(logicalDomain)
+    );
+    const partsToClear = new Set(
+      Array.from(partDomains).filter((partDomain) => allowedParts.has(partDomain))
+    );
+    const remaining = (current.logicalDirtyParts[logicalDomain] || []).filter(
+      (partDomain) => !partsToClear.has(partDomain)
+    );
 
-  const partsToClear = new Set(partDomains);
-  const remaining = (state[logicalDomain] || []).filter(
-    (partDomain) => !partsToClear.has(partDomain)
-  );
+    if (remaining.length > 0) {
+      nextLogicalDirtyParts[logicalDomain] = remaining;
+    } else {
+      delete nextLogicalDirtyParts[logicalDomain];
+    }
 
-  if (remaining.length > 0) {
-    state[logicalDomain] = remaining;
-  } else {
-    delete state[logicalDomain];
-  }
-
-  writeDirtyState(state);
+    return {
+      ...current,
+      logicalDirtyParts: nextLogicalDirtyParts,
+    };
+  });
 }
